// src/foundation/testing/src/interfaces/results/i-test-method-result-options.ts
/**
 * @license
 * Copyright (c) Noldova.
 *
 * This source code is licensed under the license found in the
 * LICENSE file in the root directory of this source tree.
 */

import type { TestDataRow } from "../../models/discovery/test-data-row.js";

export interface ITestMethodResultOptions {
  readonly testDataRow?: TestDataRow;
  readonly failure?: unknown;
  readonly skipReason?: string;
}